Validacion de endpoints /posts/* y /posts/{id}/comments/*

+++ src/Controller/CommentController.php
@@ -3,23 +3,25 @@
 namespace App\Controller;
 
 use App\Entity\Comment;
+use App\Enum\Visibility;
 use App\Repository\CommentRepository;
 use App\Repository\PostRepository;
 use App\Service\CommentManager;
 use App\Service\PostManager;
 use Psr\Log\LoggerInterface;
 use Symfony\Bridge\Doctrine\Attribute\MapEntity;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\Attribute\Route;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
-final class CommentController extends AbstractController
+final class CommentController extends ApiController
 {
     public function __construct(
+        protected ValidatorInterface $validator,
         private LoggerInterface $logger,
 
         private CommentRepository $commentRepository,
@@ -33,13 +35,35 @@ final class CommentController extends AbstractController
     ) {}
 
     #[Route('/posts/{idp}/comments/{idc}', methods: ['GET'], name: 'app_comment_get')]
-    public function get(int $idp, int $idc): JsonResponse
+    public function get(string $idp, string $idc): JsonResponse
     {
         $user = $this->getUser(); /** @var \App\Entity\User $user */
 
+        // Validación
+        $this->validate(
+            ['idp' => $idp, 'idc' => $idc],
+            new Assert\Collection([
+                'fields' => [
+                    'idp' => [
+                        new Assert\Type('digit'),
+                        new Assert\Positive(),
+                    ],
+                    'idc' => [
+                        new Assert\Type('digit'),
+                        new Assert\Positive(),
+                    ],
+                ],
+                'allowExtraFields' => true,
+            ])
+        );
+
         // Encontrar comentario
-        $comment = $this->commentRepository->findById($idc, $user);
-        if ($comment === null) throw $this->createNotFoundException();
+        $comment = $this->commentRepository->findById($idc, $idp, $user);
+        if (!$comment) {
+            return $this->json([
+                'error' => ['message' => 'Comentario no encontrado.']
+            ], 404);
+        }
 
         // Responder
         return $this->json([
@@ -48,7 +72,7 @@ final class CommentController extends AbstractController
     }
 
     #[Route('/posts/{idp}/comments', methods: ['GET'], name: 'app_comment_list')]
-    public function list(int $idp, Request $request): JsonResponse
+    public function list(string $idp, Request $request): JsonResponse
     {
         $user = $this->getUser(); /** @var \App\Entity\User $user */
 
@@ -56,8 +80,23 @@ final class CommentController extends AbstractController
         $data = $request->query->all();
         $data['post'] = $idp;
 
-        // Validación con DTO
-        // TODO
+        // Validación
+        $data = $this->validate(
+            $data,
+            new Assert\Collection([
+                'fields' => [
+                    'post' => [
+                        new Assert\Type('digit'),
+                        new Assert\Positive(),
+                    ],
+                    'page' => [
+                        new Assert\Type('digit'),
+                        new Assert\Positive(),
+                    ],
+                ],
+                'allowExtraFields' => true,
+            ])
+        );
 
         // Encontrar comentarios
         $comments = $this->commentRepository->findByFilters($data, $user);
@@ -69,7 +108,7 @@ final class CommentController extends AbstractController
     }
 
     #[Route('/posts/{idp}/comments', methods: ['POST'], name: 'app_comment_post')]
-    public function post(int $idp, Request $request): JsonResponse
+    public function post(string $idp, Request $request): JsonResponse
     {
         // Control de acceso
         $user = $this->getUser(); /** @var \App\Entity\User $user */
@@ -81,9 +120,10 @@ final class CommentController extends AbstractController
 
         // Parseo del request JSON
         $data = json_decode($request->getContent(), true);
+        $data['postId'] = $idp;
 
-        // Validacion con DTO
-        // TODO
+        // Validación
+        $data = $this->validateComment($data);
 
         // Crear comentario
         $post = $this->postRepository->find($idp);
@@ -117,11 +157,7 @@ final class CommentController extends AbstractController
     }
 
     #[Route('/posts/{idp}/comments/{idc}', methods: ['PATCH'], name: 'app_comment_patch')]
-    public function patch(
-        #[MapEntity(mapping: ['idc' => 'id'])]
-        Comment $comment,
-        Request $request
-    ): JsonResponse {
+    public function patch(string $idp, string $idc, Request $request): JsonResponse {
         // Control de acceso
         $user = $this->getUser(); /** @var \App\Entity\User $user */
         if ($user === null) {
@@ -129,14 +165,26 @@ final class CommentController extends AbstractController
                 'error' => ['message' => 'Se requiere autenticación para acceder a este endpoint.']
             ], 401);
         }
+
+        // Parseo del request JSON
+        $data = json_decode($request->getContent(), true);
+        $data['postId'] = $idp;
+
+        // Validación
+        $data = $this->validateComment($data, true);
+
+        // Encontrar comentario
+        $comment = $this->commentRepository->findById($idc, $idp, $user);
+        if (!$comment) {
+            return $this->json([
+                'error' => ['message' => 'Comentario no encontrado.']
+            ], 404);
+        }
         if ($user !== $comment->getUser() && !\in_array('ROLE_ADMIN', $user->getRoles())) {
             return $this->json([
                 'error' => ['message' => 'No tienes permisos suficientes para acceder a este endpoint.']
             ], 403);
         }
-
-        // Parseo del request JSON
-        $data = json_decode($request->getContent(), true);
 
         // Actualizar comentario
         $comment = $this->commentManager->update($data, $comment, $user);
@@ -149,10 +197,7 @@ final class CommentController extends AbstractController
     }
 
     #[Route('/posts/{idp}/comments/{idc}', methods: ['DELETE'], name: 'app_comment_delete')]
-    public function delete(
-        #[MapEntity(mapping: ['idc' => 'id'])]
-        Comment $comment
-    ): JsonResponse {
+    public function delete(string $idp, string $idc): JsonResponse {
         // Control de acceso (SOLO ADMINS)
         $user = $this->getUser(); /** @var \App\Entity\User $user */
         if ($user === null) {
@@ -162,11 +207,74 @@ final class CommentController extends AbstractController
             return $this->json(['error' => ['message' => 'No tienes permisos suficientes para acceder a este endpoint.']], 403);
         }
 
+        // Validación
+        $this->validate(
+            ['idp' => $idp, 'idc' => $idc],
+            new Assert\Collection([
+                'fields' => [
+                    'idp' => [
+                        new Assert\Type('digit'),
+                        new Assert\Positive(),
+                    ],
+                    'idc' => [
+                        new Assert\Type('digit'),
+                        new Assert\Positive(),
+                    ],
+                ],
+                'allowExtraFields' => true,
+            ])
+        );
+
+        // Encontrar comentario
+        $comment = $this->commentRepository->findById($idc, $idp, $user);
+        if (!$comment) {
+            return $this->json([
+                'error' => ['message' => 'Comentario no encontrado.']
+            ], 404);
+        }
+
         // Eliminar comentario
         $this->commentManager->delete($comment);
 
         return $this->json([
             'message' => 'Comentario eliminado.'
         ], 200);
+    }
+
+    private function validateComment(array $input, bool $patch = false)
+    {
+        $fields = [
+            'postId' => [
+                new Assert\Type('digit'),
+                new Assert\Positive(),
+            ],
+            'content' => [
+                new Assert\Type('string'),
+                new Assert\Length(max: 5000),
+            ],
+            'replyingTo' => new Assert\Optional([
+                new Assert\Sequentially([
+                    new Assert\Type(['type' => 'digit']),
+                    new Assert\Positive(),
+                ]),
+                new Assert\IsNull(),
+            ]),
+            'visibility' => [
+                new Assert\Type('string'),
+                new Assert\Choice(array_column(Visibility::cases(), 'value')),
+            ]
+        ];
+
+        if (!$patch) {
+            unset($fields['visibility']);
+        }
+
+        return $this->validate(
+            $input,
+            new Assert\Collection([
+                'fields' => $fields,
+                'allowMissingFields' => $patch,
+            ])
+        );
     }
 }
